Cplummer/ld51 46 (#23)
* Update ApiTest scene

* PlayerManager and models updates

* API model updates

* Update stats in API when game is over

+++ API/LudumDare51.API.Models/Player.cs
@@ -9,11 +9,11 @@ public class Player : BaseResource
     public string ShirtColor { get; set; }
     public string PantColor { get; set; }
     
-    public ushort TimesPlayed { get; set; }
+    public int TimesPlayed { get; set; }
     
-    public ushort Score { get; set; }
+    public int Score { get; set; }
     
-    public ushort Waves { get; set; }
+    public int Waves { get; set; }
 
     public Player() { }
